finally got primitive tokens generating with correct values.
